create BaseEvent class, from which FeaturedEvent and GridEvent inherit from
